requirements.txt
# TUI Framework
textual[syntax]>=0.47.0

# YAML parsing
pyyaml>=6.0

# Date/time utilities
python-dateutil>=2.8.2
